Added Fps Graph View and Dummy Settings Screen and Swipe-to-dismiss

+++ notxw_social/app/src/main/java/com/collabora/xwperf/notxw_social/IFpsListener.java
@@ -1,5 +1,5 @@
 package com.collabora.xwperf.notxw_social;
 
 public interface IFpsListener {
-    public void onFpsCount(int fps);
+    public void onFpsCount(long fps);
 }
